gcc: hack to stub multilib
for some reason, gcc's build system will install its shared libraries in the lib64 folder in the toolchain. it shouldn't do this, but it still does it anyways. stub out this one gcc command line flag so gcc's build system never creates or uses a lib64 folder on any arch.